% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footer.R
\name{add_footer}
\alias{add_footer}
\title{Add a footer with a caption and a logo into a ggplot}
\usage{
add_footer(x, caption = "", logo = NULL, height = 0.5)
}
\arguments{
\item{x}{a ggplot object}

\item{caption}{text to display in the footer}

\item{logo}{path to a jpg/jpeg/png file. Set to \code{NA} to disable. Defaults to
a short THL pearl logo.}
}
\description{
Add a footer with a caption and a logo into a ggplot
}
