% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_overlaps.R
\name{identify_overlaps}
\alias{identify_overlaps}
\title{Identify overlapping intervals}
\usage{
identify_overlaps(data, start, end, max_gap = 0)
}
\arguments{
\item{data}{A data frame. All variables are evaluated in this context.}

\item{start}{An expression that gives the start of the interval.}

\item{end}{An expression that gives the end of the interval.}

\item{max_gap}{Mximum distance between the star and end of consecutive
intervals that are still counted as overlapping.}
}
\value{
\code{data} with a new \code{.seq} column that creates an identifier which
groups overlapping rows together
}
\description{
Identify overlapping intervals
}
