% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{Interleave elements of vectors}
\usage{
interleave(...)
}
\arguments{
\item{...}{vectors to be interleaved}
}
\value{
result of \code{c(...)} ordered to interleave elements of \code{...}
}
\description{
Use to interleave elements in multiple vectors to a single vector
}
\examples{
x <- paste0("x", 1:5)
y <- paste0("y", 1:2)
z <- paste0("z", 1:3)
interleave(x, y, z)
}
