% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_overlaps.R
\name{merge_overlaps}
\alias{merge_overlaps}
\title{Merge overlapping intervals in a data frame}
\usage{
merge_overlaps(tbl, start, end, ..., max_gap = 0)
}
\description{
Merge overlapping intervals in a data frame
}
