% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{Calculate counts and percentages}
\usage{
np(df, x, ...)
}
\arguments{
\item{df}{a data frame}

\item{x}{a variable whose levels are counted}

\item{...}{additional variables to count by, defining also the grouping that
is used to calculate percentages}
}
\description{
Wraps \code{\link[dplyr:count]{dplyr::count()}} to calculate counts and group percentages for a given
combination of variables.
}
