% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np_gt}
\alias{np_gt}
\title{Create an np table and display it with gt}
\usage{
np_gt(df, x, ..., .fmt_p = gt::fmt_percent)
}
\arguments{
\item{df}{a data frame}

\item{x}{a variable whose levels are counted}

\item{...}{additional variables to count by, defining also the grouping that
is used to calculate percentages}

\item{.fmt_p}{a gt formatter applied to the \code{p} column}
}
\description{
Create an np table and display it with gt
}
\examples{
np_gt(mtcars, cyl)
np_gt(mtcars, cyl, am)
np_gt(mtcars, cyl, am, vs)
}
