% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-.R
\name{str_align}
\alias{str_align}
\title{Align a vector of character strings}
\usage{
str_align(x, align = c("left", "right", "center"))
}
\arguments{
\item{x}{a character vector}

\item{align}{string specifying alignment: in \code{c("left", "right", "center")}}
}
\value{
\code{x} with all elements padded to the maximum width and aligned
x <- c("short", "medium length", "and a rather long string")
str_align(x, "left")
str_align(x, "right")
str_align(x, "center")
}
\description{
Align a vector of character strings
}
