% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-.R
\name{str_align}
\alias{str_align}
\title{Align a vector of character strings}
\usage{
str_align(x, align = "left")
}
\arguments{
\item{x}{A character vector.}

\item{align}{A character vector specifying alignment. Valid values are
\code{"left"}, \code{"right"}, and \code{"center"}.}
}
\value{
\code{x} with all elements padded to the maximum width and aligned
}
\description{
Align a vector of character strings by padding appropriately with spaces.
Naturally fails hopelessly if the text is rendered in a non-monospace font.
}
\examples{
x <- c("short", "medium length", "and a rather long string")
str_align(x, "left")
str_align(x, "right")
str_align(x, "center")

# can also use a vector of alignments
str_align(x, c("center", "right", "left"))
}
