% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-.R
\name{str_align}
\alias{str_align}
\title{Align a vector of character strings}
\usage{
str_align(x, align = c("left", "right", "center"))
}
\arguments{
\item{x}{a character vector}

\item{align}{string specifying alignment}
}
\value{
\code{x} with all elements padded to the maximum width and aligned
}
\description{
Align a vector of character strings by padding approriately with spaces.
Naturally fails hopelessly if text is rendered in a non-monospace font.
}
\examples{
x <- c("short", "medium length", "and a rather long string")
str_align(x, "left")
str_align(x, "right")
str_align(x, "center")

# can also use a vector of alignments
str_align(x, c("center", "right", "left"))
}
